[CHANGE] Parser modularization
